+++ adminer/lang/en.inc.php
@@ -11,7 +11,6 @@ Lang::$translations = array(
 	'%d row(s)' => array('%d row', '%d rows'),
 	'%d item(s) have been affected.' => array('%d item has been affected.', '%d items have been affected.'),
 	'%d row(s) have been imported.' => array('%d row has been imported.', '%d rows have been imported.'),
-	'%d e-mail(s) have been sent.' => array('%d e-mail has been sent.', '%d e-mails have been sent.'),
 	'%d in total' => '%d in total',
 	'%d query(s) executed OK.' => array('%d query executed OK.', '%d queries executed OK.'),
 );