+++ adminer/lang/vi.inc.php
@@ -271,13 +271,6 @@ Lang::$translations = array(
 	'Delete' => 'Xoá',
 	'You have no privileges to update this table.' => 'Bạn không có quyền sửa bảng này.',
 
-	'E-mail' => 'Địa chỉ email',
-	'From' => 'Người gửi',
-	'Subject' => 'Chủ đề',
-	'Attachments' => 'Đính kèm',
-	'Send' => 'Gửi',
-	'%d e-mail(s) have been sent.' => '%d thư đã gửi.',
-
 	// data type descriptions
 	'Numbers' => 'Số',
 	'Date and time' => 'Ngày giờ',